data_ingestion/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
mysql-connector-python==9.1.0
fastapi==0.115.4
uvicorn==0.22.0
scikit-learn
joblib